ua_run.cc: added support for multiple jobids
